+++ noodles-cram/src/data_container/compression_header/data_series_encoding_map/data_series.rs
@@ -33,6 +33,7 @@ pub static STANDARD_DATA_SERIES: &[DataSeries; 28] = &[
     DataSeries::QualityScores,
 ];
 
+/// A CRAM data container compression header data series.
 #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
 pub enum DataSeries {
     /// BAM bit flags (`BF`).