+++ noodles-cram/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Changed
 
+  * cram/data_container/compression_header/data_series_encoding_map/data_series:
+    Increase visibility of `DataSeries`.
+
   * cram/writer: Increased visibility of module.
 
   * cram/writer/builder: Change `build` to `build_with_writer`.